+++ rewarped/environment.py
@@ -420,8 +420,7 @@ class Environment:
             with wp.ScopedTimer("render", False):
                 self.render_time += self.frame_dt
                 self.renderer.begin_frame(self.render_time)
-                # render state 1 (swapped with state 0 just before)
-                self.renderer.render(state or self.state_1)
+                self.renderer.render(state or self.state_0)
                 self.renderer.end_frame()
 
     def print_model_info(self):