+++ rewarped/envs/dexdeform/hand.py
@@ -320,14 +320,13 @@ class Hand(MPMWarpEnvMixin, WarpEnv):
             with wp.ScopedTimer("render", False):
                 self.render_time += self.frame_dt
                 self.renderer.begin_frame(self.render_time)
-                # render state 1 (swapped with state 0 just before)
-                self.renderer.render(state or self.state_1)
-                self.render_mpm(state=state)
-                # self.render_mpm_halves(state=state)
+                self.renderer.render(state or self.state_0)
+                self.render_mpm(state)
+                # self.render_mpm_halves(state)
                 self.renderer.end_frame()
 
     def render_mpm_halves(self, state=None):
-        state = state or self.state_1
+        state = state or self.state_0
 
         # render mpm particles
         particle_q = state.mpm_x