+++ rewarped/envs/softgym/transport.py
@@ -254,9 +254,8 @@ class Transport(MPMWarpEnvMixin, WarpEnv):
             with wp.ScopedTimer("render", False):
                 self.render_time += self.frame_dt
                 self.renderer.begin_frame(self.render_time)
-                # render state 1 (swapped with state 0 just before)
-                self.renderer.render(state or self.state_1)
-                self.render_mpm(state=state)
+                self.renderer.render(state or self.state_0)
+                self.render_mpm(state)
 
                 # render target
                 target_q = self.target_q + self.env_offsets